Module restructuring
Renaming of module files; move LogEntry to its own file.
